more work on running ELF binaries
- simple static and dynamically linked binaries now run successfully
- change ELF loader fix to just zero memory directly instead of performing an additional read
- add junction_run binary that execs the binary passed on the command line

TODO:
- test more complicated binaries
- implement glibc interception of relevant symbols and syscalls
- associate a Proc with the running binary
- emulate all the syscalls...

+++ scripts/install_cmake.sh
@@ -10,7 +10,9 @@ mkdir -p $BIN_DIR
 cd $BIN_DIR
 
 # Install a specific version of cmake
-wget https://github.com/Kitware/CMake/releases/download/v3.24.2/cmake-3.24.2-linux-x86_64.sh
-chmod +x cmake-3.24.2-linux-x86_64.sh
-./cmake-3.24.2-linux-x86_64.sh --skip-license
-rm cmake-3.24.2-linux-x86_64.sh
+if [ ! -f $BIN_DIR/bin/cmake ]; then
+  wget https://github.com/Kitware/CMake/releases/download/v3.24.2/cmake-3.24.2-linux-x86_64.sh
+  chmod +x cmake-3.24.2-linux-x86_64.sh
+  ./cmake-3.24.2-linux-x86_64.sh --skip-license
+  rm cmake-3.24.2-linux-x86_64.sh
+fi
